Create authapi service Clean unused code Fix create access policy response
