+++ pkg/services/cloudmigration/cloudmigration.go
@@ -3,13 +3,13 @@ package cloudmigration
 import (
 	"context"
 
-	"github.com/grafana/grafana/pkg/services/gcom"
+	"github.com/grafana/grafana/pkg/services/authapi"
 	"github.com/grafana/grafana/pkg/services/user"
 )
 
 type Service interface {
 	// GetToken Returns the cloud migration token if it exists.
-	GetToken(ctx context.Context) (gcom.TokenView, error)
+	GetToken(ctx context.Context) (authapi.TokenView, error)
 	// CreateToken Creates a cloud migration token.
 	CreateToken(ctx context.Context) (CreateAccessTokenResponse, error)
 	// ValidateToken Sends a request to GMS to test the token.
