+++ pkg/setting/setting_cloud_migration.go
@@ -8,6 +8,7 @@ import (
 type CloudMigrationSettings struct {
 	IsTarget                    bool
 	GcomAPIToken                string
+	AuthAPIUrl                  string
 	SnapshotFolder              string
 	GMSDomain                   string
 	GMSStartSnapshotTimeout     time.Duration
@@ -33,6 +34,7 @@ func (cfg *Cfg) readCloudMigrationSettings() {
 	cloudMigration := cfg.Raw.Section("cloud_migration")
 	cfg.CloudMigration.IsTarget = cloudMigration.Key("is_target").MustBool(false)
 	cfg.CloudMigration.GcomAPIToken = cloudMigration.Key("gcom_api_token").MustString("")
+	cfg.CloudMigration.AuthAPIUrl = cloudMigration.Key("auth_api_url").MustString("")
 	cfg.CloudMigration.SnapshotFolder = cloudMigration.Key("snapshot_folder").MustString("")
 	cfg.CloudMigration.GMSDomain = cloudMigration.Key("domain").MustString("")
 	cfg.CloudMigration.GMSValidateKeyTimeout = cloudMigration.Key("validate_key_timeout").MustDuration(5 * time.Second)
